removing unnecessary stuff (for now) and changing route name
